Add Styleguide
This commit adds a styleguide page to the app, which is useful
for front-end devs working on the theme as a reference point for
existing styles. Unfortunately, the first thing this styleguide
tells me is that there are almost no vanilla styles, because
everything is special one-off classes. Still, this gets us a
jumping-on point for any CSS devs and themers to start helping
out as we aim to modernize the theme and make it responsive.