Replace reset.css with normalize.css
Reset stylesheets are no longer a best practice due to heavy-handed
affects on a site. The preferred option is to use normalize.css, which
is regularly updated and works to only change things that are different
between browsers.

This commit removes reset.css and replaces it with normalize.css.
It also updated some of the site styles to avoid any visual changes as
a result of this change. Primarily, that meant zeroing out margins on
headlines and removing margins and list styles from many unordered
lists on the site.

This change should make it much easier for other front-enders to work on
MLTSHP, because markup will behave the way it's expected to. Consider
this to be shoring up the foundation of the house before we renovate it.

This commit also affects the markup for the admin pages, just to
standardize them all to use the standard page template and admin styles,
which wasn't happening before. It's not really related to the normalize
change, except that I had to edit a couple admin pages to add classes
to lists to avoid any visual changes, and it was bugging me.