Fix Jump To Previous & Bookmark Layout
This commit fixes some layout issues with the bookmark and the jump to
previous link. On small screens, they would wrap to two lines, and on
large screens there wasn't enough space between them.

This commit fixes it by reducing the amount of visible text on small
screens and increasing the space between on large screens.

fixes #201