Fix Some Layout Bugs
1) Fix a layout bug affecting the grid layout for the permalink page,
where it was trying to space the body content out to match the height
of the sidebar. Now the page content will be no taller than it should
be and the final cell in the grid will expand to match the height of
the sidebar if it's taller.

2) Fix a bug where the whitespace in the markup for the shake
description was adding whitespace in the layout because it was set to
white-space:pre-wrap to preserve the users' linebreaks.

fixes #205
fixes #221