Prevent Forcible Resizing of Quick Comment Box
This commit prevents an odd behavior where if you resized the quick
comment textarea, it would snap back to the original size. This was
definitely a bug, caused by an old bit of code to expand the textarea
when the user clicked into it. By changing `height` to `min-height`
the intended behavior is preserved, but users can make the textarea
larger if needed.

fixes #343