Fix Icon Buttons in Safari
This commit fixes a bug where the icon in an icon button on Safari was
not properly centered. This turns out to be a bug in the Safari and
Firefox implimentation of flexbox, where buttons aren't allowed to be
flex containers. Luckily, the fix was really simple. Remove flexbox.

fixes #339