Use `box-sizing:border-box`
This commit applies `box-sizing:border-box` as a universal style, and
updates any instances of fixed widths that don't work properly with
this style.

Border-box is a best practice for developer ease-of-use, to ensure that
box padding and borders are included in widths, so things like
`width:100%` will behave predictably.

fixes #400