Enable Font Features for Headlines
This commit applies a set of font feature rules to headlines to enable
the browser to use kerning and ligatures. Unfortunately, this doesn't
have any effect on Mac since the system font doesn't support it, but
it shouldn't cause any problems to have these rules in case they start
working someday.

fixes #242