Allow Shake Selector to Scroll
There's a problem you could run into if you have a lot of shakes where
the shake selector in the New Post popover scrolled offscreen and you
couldn't access the overflowed shakes.

This commit fixes that by allowing the shake selector in the new post
popover to scroll if it's longer than 66% of the viewport height.

fixes #439