Display errors from the create subscription API to the user (#493)
When a user fills in the credit card modal, the information may be
valid and indicate that it was accepted, but the card is not charged
until we send a subscription creation request along with the card
source token we got from Stripe. If the card is declined for some
reason, this request will fail. This change adds a page to explain
there was a failure and displays a message from Stripe describing
the failure. This change also sends through the payment source token
when we have one for the subscription APIs, now that we always
prompt (more to verify for cases where a card is on file already)
for payment information.