Switch Main Stylesheet to MLTSHP-Patterns v2 (#620)
* Switch Main Stylesheet to MLTSHP-Patterns v2

This PR updates the `main.scss` file to be the output of the MLTSHP
Pattern Library. With this change, the source for the CSS in in the
MLTSHP/mltshp-patterns repo, where CSS development can happen more
easily. Releases to production can happen via PRs like this.

Also adds a nice dark mode!