Additional changes to remove S3 from unit tests
Actual S3 calls are not needed for testing MLTSHP itself. This makes it
easier to run tests (for CI or locally), since it removes the need for the
'fakes3' service which had a license stipulation. The fakes3 service was
also segfaulting on Apple Silicon chips, making it impossible to run unit
tests on my laptop.

Also switching to MySQL memory tables to try to improve test performance.