Updating unit tests made for torrez.org to use https (#740)
* Updating unit tests made for torrez.org to use https

A few of our unit tests execute against a live domain. Recently, these
started failing because they no longer serve via http, but redirect to a
https address, breaking test assertions.

Eventually these tests should be replaced with something that doesn't
rely on a third-party site, but this will repair our test suite for now.

* Mock image request for PickerPopupHandler

This mocks the network requests attempted for posting an image using the
`PickerPopupHandler`, serving a response from the test fixture "1.png"
(a 1x1 PNG image) instead.

(The requests were actually issued previously (to notes.torrez.org),
which is hosted by Cloudflare, and serving a challenge response instead
of the actual image being requested.)

* Bypassing S3 for PickerPopupHandler unit tests

* Additional changes to remove S3 from unit tests

Actual S3 calls are not needed for testing MLTSHP itself. This makes it
easier to run tests (for CI or locally), since it removes the need for the
'fakes3' service which had a license stipulation. The fakes3 service was
also segfaulting on Apple Silicon chips, making it impossible to run unit
tests on my laptop.

Also switching to MySQL memory tables to try to improve test performance.