Updates to support external S3 signing (#784)
* Updates to support external S3 signing

Fastly will be able to handle the S3 signing and request for us,
so images and movies are served directly from the CDN layer instead
of our web stack.

* Cleanup around CDN stuff

Removed extra 'cdn_ssl_host' (what CDN doesn't use SSL?) option

Removed a lot of hardcoded domain names in the source (we have
configurable options for these, so should be using them)

Added a "use_fastly" option to condition special handling for Fastly
(where we now redirect to a "/s3" path that Fastly will route; if this
is not configured, we X-Accel-Redirect to the "/s3" path so Nginx will
route).

* Fixing content type parameter

* Update default charset to utf8mb4

(this should fix #761)