User admin feature (#824)
* User admin feature

CSS should probably be extracted to mltshp-patterns, as long as we
can segment it into a separate admin CSS bundle for admin-specific
components.

Inlined scripts here are also admin-specific, so could be placed
in an admin.js file, if there are things that can be shared with other
admin views. Presently, there isn't a need.

* Use background task for cascade deletion

* Isolate admin styles